requirements.txt
dotenv
openai
discord.py
